+++ DevHome.sln
@@ -58,16 +58,6 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevHome.Settings", "settings\DevHome.Settings\DevHome.Settings.csproj", "{72CE876A-7272-4460-92EB-845A45B743FB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevHome.SetupFlow.ElevatedServer", "tools\SetupFlow\DevHome.SetupFlow.ElevatedServer\DevHome.SetupFlow.ElevatedServer.csproj", "{940FD524-1AC0-4BBA-BBBE-1E4F2E797508}"
-	ProjectSection(ProjectDependencies) = postProject
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1} = {A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevHome.SetupFlow.ElevatedComponent", "tools\SetupFlow\DevHome.SetupFlow.ElevatedComponent\DevHome.SetupFlow.ElevatedComponent.csproj", "{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DevHome.SetupFlow.ElevatedComponent.Projection", "tools\SetupFlow\DevHome.SetupFlow.ElevatedComponent.Projection\DevHome.SetupFlow.ElevatedComponent.Projection.csproj", "{053AF75C-5CD8-497F-BA25-47435BD86047}"
-	ProjectSection(ProjectDependencies) = postProject
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1} = {A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}
-	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CoreWidgetProvider", "CoreWidgetProvider", "{32C0052F-10AF-48C0-A7A3-B0A1793397FD}"
 EndProject
@@ -383,38 +373,6 @@ Global
 		{940FD524-1AC0-4BBA-BBBE-1E4F2E797508}.Release|x64.Build.0 = Release|x64
 		{940FD524-1AC0-4BBA-BBBE-1E4F2E797508}.Release|x86.ActiveCfg = Release|x86
 		{940FD524-1AC0-4BBA-BBBE-1E4F2E797508}.Release|x86.Build.0 = Release|x86
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|Any CPU.Build.0 = Debug|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|arm64.ActiveCfg = Debug|ARM64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|arm64.Build.0 = Debug|ARM64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|x64.ActiveCfg = Debug|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|x64.Build.0 = Debug|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|x86.ActiveCfg = Debug|x86
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Debug|x86.Build.0 = Debug|x86
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|Any CPU.ActiveCfg = Release|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|Any CPU.Build.0 = Release|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|arm64.ActiveCfg = Release|ARM64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|arm64.Build.0 = Release|ARM64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|x64.ActiveCfg = Release|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|x64.Build.0 = Release|x64
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|x86.ActiveCfg = Release|x86
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1}.Release|x86.Build.0 = Release|x86
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|Any CPU.ActiveCfg = Debug|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|Any CPU.Build.0 = Debug|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|arm64.ActiveCfg = Debug|ARM64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|arm64.Build.0 = Debug|ARM64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|x64.ActiveCfg = Debug|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|x64.Build.0 = Debug|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|x86.ActiveCfg = Debug|x86
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Debug|x86.Build.0 = Debug|x86
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|Any CPU.ActiveCfg = Release|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|Any CPU.Build.0 = Release|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|arm64.ActiveCfg = Release|ARM64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|arm64.Build.0 = Release|ARM64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|x64.ActiveCfg = Release|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|x64.Build.0 = Release|x64
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|x86.ActiveCfg = Release|x86
-		{053AF75C-5CD8-497F-BA25-47435BD86047}.Release|x86.Build.0 = Release|x86
 		{0D879E08-99AA-4019-9D04-DEA9F7C7BFC1}.Debug|Any CPU.ActiveCfg = Debug|x64
 		{0D879E08-99AA-4019-9D04-DEA9F7C7BFC1}.Debug|Any CPU.Build.0 = Debug|x64
 		{0D879E08-99AA-4019-9D04-DEA9F7C7BFC1}.Debug|arm64.ActiveCfg = Debug|ARM64
@@ -465,8 +423,6 @@ Global
 		{C5184A1B-6C10-4346-AE64-87D2157F99D2} = {222B92B1-AC7A-409D-957B-A3851D3F41B0}
 		{6254ADB1-B6FD-4D74-AF13-40C997919178} = {222B92B1-AC7A-409D-957B-A3851D3F41B0}
 		{940FD524-1AC0-4BBA-BBBE-1E4F2E797508} = {4179A05E-37F1-46CD-9218-0889EA2BB75B}
-		{A7E5FD7B-B41A-4CAE-A45A-E686DFA8ACF1} = {4179A05E-37F1-46CD-9218-0889EA2BB75B}
-		{053AF75C-5CD8-497F-BA25-47435BD86047} = {4179A05E-37F1-46CD-9218-0889EA2BB75B}
 		{0D879E08-99AA-4019-9D04-DEA9F7C7BFC1} = {32C0052F-10AF-48C0-A7A3-B0A1793397FD}
 		{493E025E-AA0A-44A2-A2A5-6DA88329194C} = {4179A05E-37F1-46CD-9218-0889EA2BB75B}
 	EndGlobalSection
