--- tools/SetupFlow/DevHome.SetupFlow.ElevatedComponent/Helpers/ElevatedConfigureUnitTaskResult.cs
@@ -1,18 +1,0 @@
-﻿// Copyright (c) Microsoft Corporation and Contributors
-// Licensed under the MIT license.
-
-namespace DevHome.SetupFlow.ElevatedComponent.Helpers;
-
-/// <summary>
-/// Class for the result of an individual unit/resource inside <see cref="ElevatedConfigureTaskResult"/>
-/// </summary>
-public sealed class ElevatedConfigureUnitTaskResult
-{
-    public string? UnitName { get; set; }
-
-    public string? Intent { get; set; }
-
-    public bool IsSkipped { get; set; }
-
-    public int HResult { get; set; }
-}
