--- tools/SetupFlow/DevHome.SetupFlow.ElevatedComponent/NativeMethods.txt
@@ -1,33 +1,0 @@
-﻿CreateVirtualDisk
-HRESULT_FROM_WIN32
-AttachVirtualDisk
-DeviceIoControl
-CreateFile
-OpenVirtualDisk
-GetVirtualDiskPhysicalPath
-STORAGE_DEVICE_NUMBER
-IOCTL_STORAGE_GET_DEVICE_NUMBER
-PARTITION_INFORMATION_EX
-DRIVE_LAYOUT_INFORMATION_EX
-MAX_PATH
-PathIsNetworkPath
-IOCTL_STORAGE_GET_DEVICE_NUMBER_EX
-VIRTUAL_STORAGE_TYPE_DEVICE_VHDX
-VIRTUAL_STORAGE_TYPE_VENDOR_MICROSOFT 
-CREATE_DISK
-IOCTL_DISK_CREATE_DISK
-IOCTL_DISK_GET_DRIVE_LAYOUT_EX
-PARTITION_BASIC_DATA_GUID
-IOCTL_DISK_SET_DRIVE_LAYOUT_EX
-STORAGE_DEVICE_NUMBER_EX
-SetVolumeMountPoint
-DeleteVolumeMountPoint
-GetVolumePathNamesForVolumeName
-FindFirstVolume
-FindNextVolume
-FILE_FS_PERSISTENT_VOLUME_INFORMATION
-FSCTL_QUERY_PERSISTENT_VOLUME_STATE
-StrFormatByteSizeEx
-S_OK
-DetachVirtualDisk 
-OpenVirtualDisk
