+++ tools/SetupFlow/DevHome.SetupFlow.UnitTest/DevHome.SetupFlow.UnitTest.csproj
@@ -25,9 +25,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\..\src\DevHome.csproj" />
-    <ProjectReference Include="..\DevHome.SetupFlow.ElevatedComponent\DevHome.SetupFlow.ElevatedComponent.csproj">
-      <Aliases>Server</Aliases>
-    </ProjectReference>
     <ProjectReference Include="..\DevHome.SetupFlow\DevHome.SetupFlow.csproj" />
   </ItemGroup>
 </Project>
